HOOLY FUUCK
ok so...
Added changes to the links in both areas, but it works now.
Fixed teleports just being glowy obstacles, in Indus and Global. They ALL link to an empty room with a program and a server.
Todo: Fix music volume and find a music that works in the smaller rooms. And do the Fireman level.